refactor: secure api routes for sensitive data (#147)
Closes #143

+++ pages/api/modules/delete/bookmark.ts
@@ -1,7 +1,15 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
+import { getSession } from 'next-auth/react'
 import prisma from '../../../../lib/Prisma'
 
-export default async function handler( req: NextApiRequest, res: NextApiResponse) {  
+export default async function handler( req: NextApiRequest, res: NextApiResponse) { 
+  const session = await getSession({ req })
+
+  if (!session) {
+    res.status(401).json({ Error: 'UNAUTHENTICATED!' })
+    return
+  }
+   
   if (req.method === 'DELETE') {
     const deleteBookmark = await prisma.bookmark.deleteMany({
       where: {
